requirements.txt
pickle5==0.0.11
spidev==3.5
PyBluez==0.23
adafruit_circuitpython_busdevice==5.0.6
pandas==0.23.3+dfsg
pymongo==3.12.0
bluedot==2.0.0
Adafruit_Blinka==6.10.0
pynmea2==1.18.0
bluepy==1.3.0
gpiozero==1.6.2
adafruit_extended_bus==1.0.2
Adafruit_PureIO==1.1.9
board==1.0
PyQt5==5.15.4
pyserial==3.5
python_dateutil==2.8.2
smbus==1.1.post2
